requirements.txt
numpy
torch
absl_py
accelerate
gin_config
imageio
imageio[ffmpeg]
matplotlib
mediapy
opencv_contrib_python
opencv_python
Pillow
trimesh
plyfile
rawpy
ninja
scipy
scikit-image
tensorboardX
tqdm
